Bug #12319710: INVALID MEMORY READ AND/OR CRASH IN MY_UCA_CHARCMP WITH UTF32
The 5.5 version of the UTF32 collation was not enforcing the BMP range that 
it currently supports when comparing with LIKE. 
Fixed by backporting the checks for the BMP from trunk.
Added a named constant for the maximum character that can have a weight
in the weight table.
--BZR--
revision-id: georgi.kodinov@oracle.com-20110722125447-un95nb65soiiq4mn
property-branch-nick: B12319710-5.5-security
testament3-sha1: 74ff121d48465119f0ddf47d0208102b83ad8e9c