Fix for bug #13116518 - "OPEN_TABLES() SHOULD NOT ALLOCATE AND FREE NEW_FRM_MEM WITHOUT NEEDING TO".
During the process of opening tables for a statement, we allocated
memory which was used only during view loading even in cases when the
statement didn't use any views. Such an unnecessary allocation (and
corresponding freeing) might have caused significant performance
overhead in some workloads. For example, it caused up to 15% slowdown
in a simple stored routine calculating Fibonacci's numbers.

This memory was pre-allocated as part of "new_frm_mem" MEM_ROOT
initialization at the beginning of open_tables(). 

This patch addresses this issue by turning off memory pre-allocation
during initialization for this MEM_ROOT. Now, memory on this root
will be allocated only at the point when the first .FRM for a view is 
opened.

The patch doesn't contain a test case since it is hard to test the
performance improvements or the absence of memory allocation in our
test framework.
--BZR--
revision-id: dmitry.lenev@oracle.com-20111023053735-xtu0xhctx8zfae4e
property-branch-nick: mysql-5.5-13116518
testament3-sha1: 9ea5a4da1f838f91db54630f68e35063c8540b87