Bug#62712: RPM-based install of the test suite is useless for non-root users.
Although a user with read access to the mysql-test root directory
should be able to execute tests, this actually does not work in
some tests because mysqltest tries to preserve the original owner
and group when copying files. The fix is to simply not copy file
ownership information when copying files within mysqltest.