Bug#62907: mysql_plugin.test fails on a RPM-based install.
The test case for mysql_plugin has a couple of problems that make it
fail when the test is executed in the context of a RPM-based install.
One problem is the way it uses the location of the base directory from
the pathname of the mysqld binary. In a RPM-based installation, the
mysqld base directory is not an appropriate location to create
temporary files. Another problem is that the test does not cleanup by
itself. The test creates a couple of directories/files that are not
fully removed at the end of the test.

The solution is use a temporary directory as the base directory for
the test and to remove any files that are created.