BUG#13337202 - REPLICATION PERFORMANCE DROP DUE TO "THE BINLOG MAY BE CORRUPTED" FLOOD
In patch mysql-5.5:revno:3097.92.133, we made the gcc 4.6.1 compiler
to stop complaining about the fact that binlog_can_be_corrupted was
defined but not used. The fix consisted in checking the variable
and printing a warning message.

However, the fix caused a regression as a message was being printed
out when there was no corrupted binary log causing performance
problems and triggering users' suspicions when there was no need.

In BUG#13337202, we do not print any message and use the variable
in an "if" with an empty body to keep the compiler happy.
--BZR--
revision-id: alfranio.correia@oracle.com-20111102152818-gt8mqvrw3mlxbvdc
property-branch-nick: mysql-5.5.18-release
testament3-sha1: 18e8e61259a94c067e751a3bf711f650c1a1f2de