Bug#63130: CMake-based check for the presence of a system readline            library is not reliable.
The MYSQL_FIND_SYSTEM_READLINE macro, which tests whether the
Readline library is availability by trying to compile and link a
test program, does not handle the case where the Readline library
uses termcap functions, but does not link with the termcap or curses
library itself.

Also, the History library is not properly detected/used. The
<readline/history.h> file should be include in any file that uses
the History library's features. It supplies extern declarations for
all of the library's public functions and variables, and declares
all of the public data structures.

The solution is to properly use the history library header and to
link the available curses library.