MYSQL-12: Improve MySQL memory allocation, especially under NUMA
The problem is that InnoDB buffer pool memory allocations do not
follow NUMA policies strictly because pages are lazy-allocated by
the OS, and the NUMA policy applies at the time of lazy-allocation
rather than the time of reservation. This might lead to out of
memory conditions at lazy-allocation time whenever there is memory
pressure. A more robust and predictable page allocation model is
necessary to mitigate this problem.

One strategy to mitigate this problem is to force the OS to
pre-allocate the page frames in physical memory required for the
buffer pool memory region. To this end, this change implements a new
--innodb-buffer-pool-populate option that toggles the preallocation
of physical memory for the buffer pool. The preallocation strategy
is implemented by using mmap's MAP_POPULATE flag when allocating
buffer pool memory, or by pre-faulting the pages if the former does
not work.