MYSQL-10: Modules for non-functional testing: sysbench, randgen, dbt2, etc.
Introduce a randgen test suite that builds upon MTR capability to
start a server and run an external program against it, which, in
this case, is the Random Query Generator.

The tests should be invoked with the RQG_DIR variable set, which
indicates the RQG base directory where its scripts are located. The
tests use the Perl script gentest.pl to run the RQG against the
server started by MTR.

Given that output produced by the RQG is not deterministic, the
output is redirected to a log file that is created in the directory
indicated by the RQG_LOG_DIR variable. Optionally, the standard
output can also be piped to a location specified by the RQG_STDOUT
variable. The standard error stream is not redirected. Hence,
anything printed to it will cause the test to fail.

Note that a RQG run may result in warnings being printed to the
server's error log, which will cause MTR to complain and fail the
test run. If such warnings are acceptable, invoke MTR with the
--nowarnings option to turn off the scanning of log files.

See http://forge.mysql.com/wiki/RandomQueryGeneratorIntegration for
further reference.