MYSQL-16: Augment mysqld_safe to flush caches and set numa policy.
Add a mysqld_safe option to flush (sync and drop) caches, specially
the page cache, before starting a mysqld server. This is useful for
establishing a consistent and predictable behavior for normal usage
and/or benchmarking.

Also, add an option to set an memory interleave policy. This causes
mysqld memory allocations to be interleaved in round-robing fashion
over the available nodes.