Specify the repository name for Review Board.
In order to work around issues with review requests, set the repository
name to be used by Review Board. The REPOSITORY setting can be used to
specify the path, numeric ID or name of the repository to use. In addition,
the .reviewboardrc file is a generic place for configuring a repository.