MYSQL-22: Make population of column source metadata in result sets optional
The field descriptor packets of the MySQL protocol (also know as
result set metadata), one for each selected column of the result
set, contains a set of schema object names that can significantly
increase the amount of data to be transmitted as part of a result
set. This is a problem especially for single-row queries (e.g. WHERE
primary_key=id) on a single wide-table (e.g. 50 columns), where
information such as database, table and column names will be
transmitted for each selected column.

In order to reduce the amount of data to be transferred as part of a
result set, some of these names might be emptied if the client does
not need them. This change introduces a new protocol mode that
instructs MySQL to empty certain object names that are part of the
result set metadata. Only the absolutely minimal (or essential) set
of names is preserved. If the minimal option is set, the database,
table (both original and aliased) and column names in the result set
metadata will be empty strings. Only the column alias name is
preserved.

The concept of protocol modes is similar to that of SQL modes, but
affecting the MySQL protocol instead of SQL syntax. The protocol
mode can be changed at runtime by setting the global or session
value of protocol_mode variable. Currently, the only supported
protocol mode value is MINIMAL_OBJECT_NAMES_IN_RESULT_SET_METADATA,
which causes a few object names in the result set metadata to be
empty.

The minimal option should be used with care as it constitutes a
protocol violation that might affect certain connectors. For
example, the minimal option makes the JDBC Driver for MySQL
(Connector/J) remap fields of type BLOB to VARCHAR as it considers
fields with an empty table name as originating from functions. Also,
certain methods of JDBC classes such as ResultSetMetaData and others
that represent meta-information might not yield meaningful results
if the minimal option is set.