Bug#61575: InnoDB: Error: io_setup() failed
The problem is that if Linux's native AIO support cannot be
initialized via io_setup(), InnoDB crashes. io_setup() can fail
if the specified number of events exceeds the user's limit of
available events.

The solution is to fail InnoDB initialization in a controlled
manner if io_setup() returns an error.