MYSQL-10: Modules for non-functional testing: sysbench, randgen, dbt2, etc.
Instead of executing each test case individually as a mean of having
a clean var directory, use --force-restart to start a new server and
recreate a new var directory for each test case. Also, disable the
parallel execution of test cases and the combinations file, both
will be evaluated later.