Bug#63623: Test case partition_binlog_stmt creates a file in an improper location.
Test case partition_binlog_stmt creates a file in the source
code/root mysql-test directory, which is an improper location
because mysqltest might not have write permission to the root
mysql-test directory, or because it might conflict with other
test runs in case of out-of-source builds.

The solution is to create the file in the mysqltest temporary
directory.