BUG#12969301 : mysql_plugin: enable is ignored if plugin exists
This patch changes the mechanism by which the client enables a 
plugin. Instead of using INSERT IGNORE to reload a plugin library,
it now uses REPLACE INTO. This allows users to load a library
multiple times replacing the existing values in the mysql.plugin
table. This allows users to replace the symbol reference to a
different dl name in the table. Thus permitting enabling of 
multiple versions of the same library without first disabling
the old version.

A regression test was added to ensure this feature works.

--BZR--
revision-id: chuck.bell@oracle.com-20111215211648-qffrge9yc8oh7otg
property-branch-nick: mysql-12969301
testament3-sha1: 48f55590858f3c62be5da39949f1cd0994414a04