MYSQL-31: Implement super_only mode to restrict all queries to SUPER-user only
Introduce a super_only option that is an analogue to read_only that
restricts all statements (rather than just update statements) to
users holding the SUPER privilege. That is, non-super users won't be
able to execute regular statements if the super_only option is
enabled. This is useful for administratively blocking the server
with nothing more than SQL access.