Bug#11754150: A TEST CASE FOR BUG#6857 IS DISABLED IN SP.TEST
The time comparison using current_time() stored in an int variable was giving wrong results as 
the current_time() format as an int implementation has been changed in mysql-trunk but not in mysql-5.5.

The time is stored in the format hh:mm:ss as 'time' datatype.But as an int, it is stored as hhmmss,
but only on the trunk. On mysql-5.5,as an int, it is stored as hh.
Hence, the current_time() function has been changed to unix_timestamp() function.
--BZR--
revision-id: sneha.modi@oracle.com-20111221094455-b6gr53xaqs5t92ah
property-branch-nick: mysql-5.5
testament3-sha1: 7bf3c6eee421064e3ff8f0352fd7a1357f3adfc5