Fix for bug#11758931 - 51196: SLAVE SQL: GOT AN ERROR WRITING COMMUNICATION PACKETS, ERROR_CODE: 1160
If idle FEDERATED table is evicted from the table cache when 
a connection to remote server is lost, query that initiated 
eviction may fail.
If this query is executed by slave SQL thread it may fail as well.

An error of close was stored in diagnostics area, which was later
attributed to the statement that caused eviction.

With this patch FEDERATED clears an error of close.

--BZR--
revision-id: ramil@mysql.com-20111223132248-827w1gp4f80tpjvr
property-branch-nick: mysql-5.1
testament3-sha1: a5fc8190206b091b42f17c705408bacc50b21bd5