MYSQL-30: Research safer buffer pool dump/restore method
Introduce a ENGINE_CONTROL(storage_engine, command) function
that manipulates the underlying parameters of storage engines.
In particular, internal characteristics of storage engines may
be controlled with ENGINE_CONTROL() requests. The first argument
is the name of the storage engine to be manipulated and the second
argument is a storage engine-specific command. Each command may
take a variable number of arguments.

Also, this change introduces a "prefetch_pages" command to the
InnoDB storage engine. This command may be use to prefetch
(retrieve from disk) index and data pages into the buffer pool.
The command synchronously waits for pages to be read and returns
the number of pages read. The goal of prefetching is to simulate
an initial warm-up period to populate the buffer pool and other
structures.

The prefetch command syntax is:

ENGINE_CONTROL(InnoDB, prefetch_pages, space, page [, page ...])