Bug #11754014: 45549: udf plugin_dir path separator inconsistency   and cryptic error 1126 message
The problem was that dlopen() related code was using just a subset 
of the path normalization routines used in other places.
Fixed the expansion of the pre-dlopen() behavior for plugins and UDFs
to use a platform-dependent consistent encoding of the paths.
Fixed the error dlopen() error handling to take the correct error message
and strip off the trailing newline character(s).
Fixed tests to do a platform independent replace of directories and to 
account for the traling slash.
--BZR--
revision-id: georgi.kodinov@oracle.com-20120116100428-w67nrlas2q453ra3
property-branch-nick: B11754014-5.5-unix
testament3-sha1: 0a99eeef75653cc155cb0de1ae1498f18ab22348