MYSQL-38: Counters in innodb_buffer_pool_stats seem broken
The problem was that the value of the NUMBER_PAGES_CREATED and
NUMBER_PAGES_WRITTEN columns of the INNODB_BUFFER_POOL_STATS were
actually being set to the rate of pages created and written. Also,
the NUMBER_PAGES_GET column wasn't being set at all.

The solution is to properly set the values of the PAGES_CREATE_RATE
and PAGES_WRITTEN_RATE columns and to also set a value for the
NUMBER_PAGES_GET column.