MYSQL-39: Accessing innodb_buffer_page is very slow
The problem stems from the use of MyISAM for the temporary table
that is created. Mainly that such internal temporary table is
created using the static (fixed-length) storage format, which
causes CHAR and VARCHAR columns to be padded to the column width.

Incidentally, upstream just fixed this in BUG#13350136 (TEMPORARY
MYISAM TABLES DO NOT PACK LONG VARCHAR COLUMNS). But, the fix
does not take into account the case where the temporary table is
being created for an information schema table. Normally, the
HA_OPTION_PACK_RECORD flag is set when a field for the temp table
is being created (create_tmp_field, etc.), but creating fields
for a information schema table uses a different code path
(create_tmp_field_for_schema) that does not set the flag.

The upstream fix is back ported but with a tweak to allow the
packed format to be used by temporary tables that are created for
information schema.

Original historical log:

BUG#13350136 - TEMPORARY MYISAM TABLES DO NOT PACK LONG
               VARCHAR COLUMNS

MyISAM has three record formats (fixed, packed and compressed)
out of which fixed and packed are applicable for temporary
tables. The difference is that with fixed length record format,
VARCHAR trailing bytes are written whereas for packed record
format the trailing bytes are not written. The packed record
therefore saves space at the cost of some additional memcopies.

The packed record format is currently not in use by temporary
tables, although old commit messages suggests that this has not
always been the case, e.g:

    revision: sp1r-monty@stripped
    Date: 2004-12-06
    "Internal temporary files can now use fixed length tables
    if the used VARCHAR columns are short"

This CS enables packed record format for temporary tables.

The CS also includes a fix for a MyISAM bug that is only a
problem when the packed format and a special unique constraint
(that checksums wide columns) is used. This unique constraint
only applies to internal temporary tables when uniqueness is
required on columns with a total size bigger than normal
indexes allow (1000 bytes). The problem:

A column's value in MyISAM can be read from or written to the
row buffer using MI_COLUMNDEF::offset ("normal" column
definition) or HA_KEYSEG::start (key definition). The former
is calculated in mi_open() and is the sum of the length of all
preceding columns. The latter is calculated in mi_create().
The bug was that in mi_create(), HA_KEYSEG::start was set to
the offset of the column for the on-disk format (reclength)
whereas it should have been the offset for the in-memory format.

Due to this bug, unique constraint values were written to one
offset (HA_KEYSEG::start in mi_check_unique()) and read from
another offset (MI_COLUMNDEF::offset in _mi_pack_record et al),
resulting in duplicate rows in the result set because equally
valued rows got seemingly different checksums.