MYSQL-37: Export InnoDB LSN information
Add status variables that show the current log sequence number
(Innodb_lsn_current), how far the log has been flushed to disk
(Innodb_lsn_flushed), and the position at which InnoDB last took
a checkpoint (Innodb_lsn_checkpoint).

No locking is used as the reads of the values are likely to be
atomic (if properly aligned). If that's not the case, it's not a
problem anyway.