MYSQL-39: Accessing innodb_buffer_page is very slow
When a MyISAM table uses the dynamic storage format, each row is
preceded by a bitmap that indicates which columns contain the
empty string (for string columns) or zero (for numeric columns).
If a string column has a length of zero after trailing space
removal, or a numeric column has a value of zero, the column can
just be marked in the bitmap and not saved to disk.