MYSQL-40: RPM "Obsoletes:" lines make it impossible to "yum install"           specific old versions
The problem is that the "obsoletes" property of a newer MySQL-server
package interferes with the installation of older (obsoleted) MySQL
packages from different repositories. For example, yum will replace
a MySQL-server-community package with a MySQL-server package if one
of the repositories contains a MySQL-server package that obsoletes
the MySQL-server-community package.

In order to avoid any obsoletes processing logic, the property is
simply removed. Since upgrades across major MySQL versions is not
supported, this should have no significant impact.