MYSQL-42: Allow pre-allocation of InnoDB file-per-table (ibd) tablespace files
Add the ability to specific through MIN_PAGES the initial size, in
pages, of per-table tablespaces. The option is only meaningful if
innodb_file_per_table is enabled. If the initial size is not large
enough to contain required metadata, the tablespace will be extended
to the required size. If the MIN_PAGES option is not specified,
InnoDB uses the minimum size possible.

In order to retain forward compatibility in FRMs, MIN_PAGES is
implemented as an alias for MIN_ROWS. Using MIN_ROWS has the same
effect as MIN_PAGES. If the MIN_PAGES option is used in a CREATE
TABLE statement that is logged to the binary log, the MIN_PAGES
option is also used in the statement that is sent to slaves.