MYSQL-42: Allow pre-allocation of InnoDB file-per-table (ibd) tablespace files
Add the ability to extend the tablespace of a table to be at least X
pages long if it's smaller. If the tablespace is large enough already,
nothing changes.

The minimum number of pages can be specified using the table option
MIN_PAGES in an ALTER TABLE statement. Extending the tablespace is
an in-place operation that does not require a temporary table.