MYSQL-45: InnoDB should allow flushing to be gated by age of pages
Currently using innodb_max_dirty_pages_pct especially with a large
innodb_io_capacity does not enforce any lower limit on the number of
dirty pages, and this can increase the number of writes dramatically.
While normally this would be OK (flush if you can) this is not
desirable behavior on SSDs where write lifetime is limited.

In order to reduce and control the frequency of write requests to
SSD, this change introduces a new variable named
innodb_flush_dirty_pages_age which can be used to set the minimum age
of dirty pages to be flushed from the buffer pool if the number of
dirty pages is below the maximum percentage. This impacts the
frequency with which frequently modified pages are flushed, which can
help reduce IO load in systems with frequently modified pages (hot
spots).