MYSQL-15: Implement server-side statement timeout support
Introduce an implementation of per-thread timers. The API can be used
to create one-shot timers that use a monotonically increasing clock
as the timing base. A timer expires in a given amount of time (in
milliseconds) from when the timer is armed. Upon timer expiration, a
given callback function is invoked in a separate thread. Timers can
also be cancelled (disarmed) and provide an indication of the state
(signaled or non-signaled) of the timer at the time of cancellation.

The timer API is currently only implemented on Linux and Mac OS X.